Users and Venues sorted

+++ public/js/database.js
@@ -1,5 +1,9 @@
 $(document).ready(function () {
   usersList();
+  venuesList();
+
+  // take focus away
+  document.addEventListener('click', function(e) { if(document.activeElement.toString() == '[object HTMLButtonElement]'){ document.activeElement.blur(); } });
 });
 
 // Get all Users to display
@@ -15,7 +19,8 @@ usersList = () => {
       "Authorization": `Bearer ${token}`,
     },
     success: function (users) {
-     userListSuccess(users);
+      users.sort(sort_by('name', false, function(a){return a.toUpperCase()}));
+      userListSuccess(users);
     },
     error: function (request, message, error) {
       handleException(request, message, error);
@@ -36,7 +41,7 @@ userListSuccess = (users) => {
 userAddRow = (user) => {
   // First check if a <tbody> tag exists, add one if not
   if ($("#userTable tbody").length == 0) {
-    $("#userTable").append("<tbody></tbody>");
+    $("#userTable").append("<tbody class='scrolltable'></tbody>");
   }
 
   // Append row to <table>
@@ -66,7 +71,7 @@ function userBuildTableRow(user) {
         "<span class='glyphicon glyphicon-remove' />" +
         "</button>" +
       "</td>" +
-    "</tr>";
+    "</tr>" 
 
   return ret;
 }
@@ -80,7 +85,7 @@ function userGet(ctl) {
   console.log(id);
   
   // Store product id in hidden field
-  $("#userid").val(id);
+  $("#storeid").val(id);
 
   // Call Web API to get a Product
   $.ajax({
@@ -94,7 +99,7 @@ function userGet(ctl) {
       userToFields(user);
 
       // Change Update Button Text
-      $("#updateButton").text("Update");
+      $("#userUpdateButton").text("Update");
     },
     error: function (request, message, error) {
       handleException(request, message, error);
@@ -106,19 +111,20 @@ function userToFields(user) {
   $("#username").val(user.name);
   $("#email").val(user.email);
   $("#phone").val(user.phone);
+  userOpenForm()
 }
 
 // Handle click event on Update button
-function updateClick() {
+function userUpdateClick() {
       // Build user object from inputs
       User = new Object();
-      User._id = $("#userid").val();
+      User._id = $("#storeid").val();
       User.name = $("#username").val();
       User.email = $("#email").val();
       User.phone = $("#phone").val();
       User.password = $("#password").val();
 
-      if ($("#updateButton").text().trim() == "Add") {
+      if ($("#userUpdateButton").text().trim() == "Add") {
         userAdd(User);
       }
       else {
@@ -205,9 +211,8 @@ function userAdd(user) {
 }
 
 function userAddSuccess(user) {
-  console.log(user)
   userAddRow(user);
-  formClear();
+  userFormClear();
 }
 
 // Update user in <table>
@@ -221,45 +226,42 @@ function userUpdateInTable(user) {
   $(row).remove();
 
   // Clear form fields
-  formClear();
+  userFormClear();
 
   // Change Update Button Text
-  $("#updateButton").text("Add");
+  $("#venueUpdateButton").text("Add");
 }
 
-// Handle click event on Add button
-function addClick() {
-  console.log('addClick');
-  
-}
 
 
  // Delete user from <table>
  function userDelete(ctl) {
+  $(this).blur();
+  if (confirm("Are you sure ?")){
+    var token = window.localStorage.getItem('token');
 
-  var token = window.localStorage.getItem('token');
+    var id = $(ctl).data("id");
 
-  var id = $(ctl).data("id");
-
-  // Call Web API to delete a user
-  $.ajax({
-    "url": `http://127.0.0.1:3001/users/${id}`,
-    "method": "DELETE",
-    "headers": {
-      "Content-Type": "application/json",
-      "Authorization": `Bearer ${token}`,
+    // Call Web API to delete a user
+    $.ajax({
+      "url": `http://127.0.0.1:3001/users/${id}`,
+      "method": "DELETE",
+      "headers": {
+        "Content-Type": "application/json",
+        "Authorization": `Bearer ${token}`,
+        },
+      "processData": false,
+      "data": ''
+      ,
+      success: function (user) {
+        $(ctl).parents("tr").remove();
       },
-    "processData": false,
-    "data": ''
-    ,
-    success: function (user) {
-      $(ctl).parents("tr").remove();
-    },
-    error: function (request, message, error) {
-      handleException(request, message, error);
-    }
-  });
-}
+      error: function (request, message, error) {
+        handleException(request, message, error);
+      }
+    });
+  }
+ }
 
 // Handle exceptions from AJAX calls
 handleException = (request, message, error) => {
@@ -275,9 +277,320 @@ handleException = (request, message, error) => {
 }
 
 // Clear form fields
-function formClear() {
+function userFormClear() {
   $("#username").val("");
   $("#email").val("");
   $("#phone").val("");
   $("#password").val("");
+  userCloseForm()
+}
+
+function userOpenForm() {
+  document.getElementById("userForm").style.display = "block";
+  $("#username").focus()
+}
+
+function userCloseForm() {
+  document.getElementById("userForm").style.display = "none";
+}
+
+// Get all Venues to display
+venuesList = () => {
+
+  var token = window.localStorage.getItem('token');
+
+  $.ajax({
+    "url": 'http://127.0.0.1:3001/venues',
+    "method": "GET",
+    "headers": {
+      "Content-Type": "application/json",
+      "Authorization": `Bearer ${token}`,
+    },
+    success: function (venues) {
+      venues.sort(sort_by('address', false, function(a){return a.toUpperCase()}));
+      venueListSuccess(venues);
+    },
+    error: function (request, message, error) {
+      handleException(request, message, error);
+    }
+  });
+}
+
+// Display all Venues returned from Web API call
+venueListSuccess = (venues) => {
+  // Iterate over the collection of data
+  $.each(venues, (index, venue) => {
+    // Add a row to the Venues table
+    venueAddRow(venue);
+  });
+}
+
+// Add Venue row to <table>
+venueAddRow = (venue) => {
+  // First check if a <tbody> tag exists, add one if not
+  if ($("#venueTable tbody").length == 0) {
+    $("#venueTable").append("<tbody></tbody>");
+  }
+
+  // Append row to <table>
+  $("#venueTable tbody").append(
+    venueBuildTableRow(venue));
+}
+
+// Build a <tr> for a row of table data
+function venueBuildTableRow(venue) {
+  var ret = "<tr>" +
+      "<td>" +
+        "<button type='button' " +
+          "onclick='venueGet(this);' " +
+          "class='btn btn-default' " +
+          "data-id='" + venue._id + "'>" +
+          "<span class='glyphicon glyphicon-edit' />"
+          + "</button>" +
+        "</td >" +
+        "<td>" + venue.address + "</td>" +
+        "<td>" + venue.contact.name + "</td>" +
+        "<td>" + venue.seats + "</td>" +
+        "<td>" +
+        "<button type='button' " +
+        "onclick='venueDelete(this);' " +
+        "class='btn btn-default' " +
+        "data-id='" + venue._id + "'>" +
+        "<span class='glyphicon glyphicon-remove' />" +
+        "</button>" +
+      "</td>" +
+    "</tr>" 
+
+  return ret;
+}
+
+function venueGet(ctl) {
+
+  var token = window.localStorage.getItem('token');
+
+  // Get product id from data- attribute
+  var id = $(ctl).data("id");
+  console.log(id);
+  
+  // Store product id in hidden field
+  $("#storeid").val(id);
+
+  // Call Web API to get a Product
+  $.ajax({
+    "url": `http://127.0.0.1:3001/venues/${id}`,
+    "method": "GET",
+    "headers": {
+      "Content-Type": "application/json",
+      "Authorization": `Bearer ${token}`,
+    },
+    success: function (venue) {
+      venueToFields(venue);
+
+      // Change Update Button Text
+      $("#venueUpdateButton").text("Update");
+    },
+    error: function (request, message, error) {
+      handleException(request, message, error);
+    }
+  });
+}
+
+function venueToFields(venue) {
+  $("#address").val(venue.address);
+  $("#cName").val(venue.contact.name);
+  $("#cEmail").val(venue.contact.email);
+  $("#cPhone").val(venue.contact.phone);
+  $("#seats").val(venue.seats);
+  venueOpenForm()
+}
+
+// Handle click event on Update button
+function venueUpdateClick() {
+       // Build venue object from inputs
+      Venue = new Object();
+      Contact = new Object();
+      Venue._id = $("#storeid").val();
+      Venue.address = $("#address").val();
+      Venue.contact = Contact;
+      Venue.contact.name = $("#cName").val();
+      Venue.contact.email= $("#cEmail").val();
+      Venue.contact.phone = $("#cPhone").val();
+      Venue.seats = $("#seats").val();
+
+      if ($("#venueUpdateButton").text().trim() == "Add") {
+        venueAdd(Venue);
+      }
+      else {
+        venueUpdate(Venue);
+      }
+    }
+
+
+function venueUpdate(venue) {
+
+  var token = window.localStorage.getItem('token');
+
+  data = new Object();
+  data.address = venue.address
+  data.contact = venue.contact
+  data.seats = venue.seats
+
+
+  // Call Web API to update venue
+  $.ajax({
+    "url": `http://127.0.0.1:3001/venues/${venue._id}`,
+    "method": "PATCH",
+    "headers": {
+      "Content-Type": "application/json",
+      "Authorization": `Bearer ${token}`,
+      },
+    "processData": false,
+    "data": `${JSON.stringify(data)}`
+    ,
+    success: function (venue) {
+    venueUpdateSuccess(venue);
+    },
+    error: function (request, message, error) {
+      handleException(request, message, error);
+    }
+  });
+}
+
+function venueUpdateSuccess(venue) {
+  venueUpdateInTable(venue);
+}
+
+function venueAdd(venue) {
+
+  var token = window.localStorage.getItem('token');
+
+  data = new Object();
+  data.address = venue.address
+  data.contact = venue.contact
+  data.seats = venue.seats
+
+
+  // Call Web API to add a new venue
+  $.ajax({
+    "url": `http://127.0.0.1:3001/venues`,
+    "method": "POST",
+    "headers": {
+      "Content-Type": "application/json",
+      "Authorization": `Bearer ${token}`,
+      },
+    "processData": false,
+    "data": `${JSON.stringify(data)}`
+    ,
+    success: function (venue) {
+      venueAddSuccess(venue);
+    },
+    error: function (request, message, error) {
+      handleException(request, message, error);
+    }
+  });
+}
+
+function venueAddSuccess(venue) {
+  venueAddRow(venue);
+  venueFormClear();
+}
+
+// Update venue in <table>
+function venueUpdateInTable(venue) {
+  // Find Venue in <table>
+  var row = $("#venueTable button[data-id='" + venue._id + "']")
+            .parents("tr")[0];
+  // Add changed venue to table
+  $(row).after(venueBuildTableRow(venue));
+  // Remove original product
+  $(row).remove();
+
+  // Clear form fields
+  venueFormClear();
+
+  // Change Update Button Text
+  $("#venueUpdateButton").text("Add");
+}
+
+// Handle click event on Add button
+function addClick() {
+  console.log('addClick');
+  
+}
+
+
+ // Delete venue from <table>
+ function venueDelete(ctl) {
+  $(this).blur();
+  if (confirm("Are you sure ?")){
+    var token = window.localStorage.getItem('token');
+
+    var id = $(ctl).data("id");
+
+    // Call Web API to delete a venue
+    $.ajax({
+      "url": `http://127.0.0.1:3001/venues/${id}`,
+      "method": "DELETE",
+      "headers": {
+        "Content-Type": "application/json",
+        "Authorization": `Bearer ${token}`,
+        },
+      "processData": false,
+      "data": ''
+      ,
+      success: function (venue) {
+        $(ctl).parents("tr").remove();
+      },
+      error: function (request, message, error) {
+        handleException(request, message, error);
+      }
+    });
+  }
+ }
+
+// Handle exceptions from AJAX calls
+handleException = (request, message, error) => {
+  var msg = "";
+
+  msg += "Code: " + request.status + "\n";
+  msg += "Text: " + request.statusText + "\n";
+  if (request.responseJSON != null) {
+    msg += "Message" + request.responseJSON.Message + "\n";
+  }
+
+  alert(msg);
+}
+
+// Clear form fields
+function venueFormClear() {
+  $("#address").val("");
+  $("#cName").val("");
+  $("#cEmail").val("");
+  $("#cPhone").val("");
+  $("#seats").val("");
+  
+  venueCloseForm()
+}
+
+function venueOpenForm() {
+  document.getElementById("venueForm").style.display = "block";
+  $("#address").focus()
+}
+
+function venueCloseForm() {
+  document.getElementById("venueForm").style.display = "none";
+}
+
+// sort function
+var sort_by = function(field, reverse, primer){
+
+  var key = primer ? 
+      function(x) {return primer(x[field])} : 
+      function(x) {return x[field]};
+
+  reverse = !reverse ? 1 : -1;
+
+  return function (a, b) {
+      return a = key(a), b = key(b), reverse * ((a > b) - (b > a));
+    } 
 }
